refactor(core,format): relocate the Datum value tree to gantz_core
Move the Datum enum, DatumError, to_datum/from_datum and the full serde
Serializer/Deserializer bridge from gantz_format to a new
gantz_core::datum module. The Steel-text rendering and parsing
(datum_text/datum_from_expr and helpers) stay in gantz_format, operating
on the relocated type. gantz_format re-exports Datum, DatumError,
from_datum, node_datum and to_datum at its root so downstream users are
untouched.

Motivation: the upcoming Ref extension slot (#304 Part 3, #295) stores
domain data on gantz_core::node::Ref as canonical datums, so core needs
the self-describing value type. It is also the representation a future
state-persistence feature will want at the core layer.

Pure move - no behavioural change (full suite passes unchanged).
