+++ crates/gantz_core/src/lib.rs
@@ -7,6 +7,7 @@ pub use steel;
 pub mod args;
 pub mod builtin;
 pub mod compile;
+pub mod datum;
 pub mod diagnostic;
 pub mod edge;
 pub mod graph;
